refactor(optimizer): Split table_elimination.rs into focused modules
Split the 1,708-line table_elimination.rs into focused modules under
the optimizer/table_elimination/ directory:

- mod.rs (317 lines): Main entry point and orchestration
- types.rs (18 lines): TableInfo and EliminatedTable structs
- select_analysis.rs (258 lines): SELECT list analysis functions
- join_analysis.rs (135 lines): Equijoin detection
- predicate.rs (133 lines): Local predicate extraction
- prefix.rs (208 lines): Column prefix inference (TPC-DS style)
- transform.rs (164 lines): AST transformation helpers
- tests.rs (551 lines): All test modules

All modules are under 500 lines. No functional changes - all 30
existing tests pass. Environment variable controls (TABLE_ELIM_DISABLED,
TABLE_ELIM_VERBOSE) still work.

Closes #4340
